requirements.txt
construct==2.9.45
hurry.filesize==0.9
kaitaistruct==0.8
tabulate==0.8.6
